git-svn-id: https://s1.saphirion.com/svn/rm-asset/trunk/r3@5332 0b15d6fc-ce89-4483-8625-0b0fabc32e25

+++ src/core/u-bmp.c
@@ -173,7 +173,7 @@ void Map_Bytes(void *dstp, REBYTE **srcp, char *map) {
 
 		case 'l':
 			if (longaligned()) {
-				while(((unsigned long)dst)&3)
+				while(((uintptr_t)dst)&3)
 					dst++;
 			}
 			*((REBCNT *)dst) = *((REBCNT *)src);
@@ -230,7 +230,7 @@ void Unmap_Bytes(void *srcp, REBYTE **dstp, char *map) {
 
 		case 'l':
 			if (longaligned()) {
-				while(((unsigned long)src)&3)
+				while(((uintptr_t)src)&3)
 					src++;
 			}
 			*((REBCNT *)dst) = *((REBCNT *)src);
